Enhanced Mechanics
I tried to touch on some radius which the robots can see their targets
but I think I'll elt that go for now, seeing that it makes the game run
significantly slower.
Other than that this update improves on how robots drop their equipment
on death. Perhaps next update will focus on guns and other firearms.
That should be fun.